+++ components/airlink_web/include/airlink_web.h
@@ -4,3 +4,4 @@
 #include "esp_err.h"
 
 esp_err_t airlink_web_start(bool recovery_mode, bool read_only_mode);
+bool airlink_web_ready(void);